Split BookedSlot into own module
